feat(cli): make sanity new own project scaffolding
