+++ packages/@sanity/cli/src/actions/scaffold/__tests__/scaffoldProject.test.ts
@@ -1,30 +1,26 @@
+import {mkdtemp, rm, writeFile} from 'node:fs/promises'
+import {tmpdir} from 'node:os'
 import path from 'node:path'
 
 import {beforeEach, describe, expect, test, vi} from 'vitest'
 
-import {
-  FRONTEND_ENV_PREFIX_OVERRIDES,
-  manualScaffoldCommands,
-  scaffoldProject,
-} from '../scaffoldProject.js'
+import {isStudioScaffoldTargetAvailable, scaffoldProject} from '../scaffoldProject.js'
 
+const mockWriteFileSync = vi.hoisted(() => vi.fn())
 const mockDetectFrameworkRecord = vi.hoisted(() => vi.fn())
 const mockDirIsEmptyOrNonExistent = vi.hoisted(() => vi.fn())
-const mockAppendEnvValues = vi.hoisted(() => vi.fn())
 const mockGetProjectDefaults = vi.hoisted(() => vi.fn())
 const mockInitStudio = vi.hoisted(() => vi.fn())
 const mockResolvePackageManager = vi.hoisted(() => vi.fn())
 const mockCreateFrontend = vi.hoisted(() => vi.fn())
 const mockInstallFrontendDeps = vi.hoisted(() => vi.fn())
 
+vi.mock('node:fs', () => ({writeFileSync: mockWriteFileSync}))
 vi.mock('../../../util/detectFramework.js', () => ({
   detectFrameworkRecord: mockDetectFrameworkRecord,
 }))
 vi.mock('../../../util/dirIsEmptyOrNonExistent.js', () => ({
   dirIsEmptyOrNonExistent: mockDirIsEmptyOrNonExistent,
-}))
-vi.mock('../../../util/envFile.js', () => ({
-  appendEnvValues: mockAppendEnvValues,
 }))
 vi.mock('../../../util/getProjectDefaults.js', () => ({
   getProjectDefaults: mockGetProjectDefaults,
@@ -56,16 +52,18 @@ const options = {
   token: 'sk-token',
   workDir: '/tmp/project',
 }
+const envContents = [
+  '# Added by `sanity new`. Keep this file out of git: it holds a live project token.',
+  'SANITY_AUTH_TOKEN="sk-token"',
+  'SANITY_DATASET="production"',
+  'SANITY_PROJECT_ID="abc123"',
+  '',
+].join('\n')
 
 beforeEach(() => {
   vi.clearAllMocks()
   mockDetectFrameworkRecord.mockResolvedValue(undefined)
   mockDirIsEmptyOrNonExistent.mockResolvedValue(true)
-  mockAppendEnvValues.mockReturnValue({
-    created: true,
-    skippedKeys: [],
-    wroteKeys: ['value'],
-  })
   mockGetProjectDefaults.mockResolvedValue({projectName: 'My Project'})
   mockInitStudio.mockResolvedValue(undefined)
   mockResolvePackageManager.mockResolvedValue('npm')
@@ -73,37 +71,34 @@ beforeEach(() => {
   mockInstallFrontendDeps.mockResolvedValue(undefined)
 })
 
-describe('manualScaffoldCommands', () => {
-  test('provides unattended Studio and frontend recovery commands', () => {
-    expect(manualScaffoldCommands({dataset: 'production', projectId: 'abc123'})).toEqual([
-      'npx sanity init --project abc123 --dataset production --output-path sanity --no-mcp --no-skills --no-git -y',
-      expect.stringContaining('create-next-app@^16 web'),
-    ])
-  })
-})
-
 describe('scaffoldProject', () => {
+  test('treats an existing Studio file as unavailable', async () => {
+    const workDir = await mkdtemp(path.join(tmpdir(), 'sanity-new-'))
+
+    try {
+      await writeFile(path.join(workDir, 'sanity'), 'not a directory')
+
+      await expect(isStudioScaffoldTargetAvailable(workDir)).resolves.toBe(false)
+      expect(mockDirIsEmptyOrNonExistent).not.toHaveBeenCalled()
+    } finally {
+      await rm(workDir, {force: true, recursive: true})
+    }
+  })
+
   test('refuses a non-empty Studio target before writing files', async () => {
     mockDirIsEmptyOrNonExistent.mockResolvedValue(false)
 
     await expect(scaffoldProject(options)).rejects.toThrow('./sanity is not empty')
 
     expect(mockInitStudio).not.toHaveBeenCalled()
-    expect(mockAppendEnvValues).not.toHaveBeenCalled()
+    expect(mockWriteFileSync).not.toHaveBeenCalled()
   })
 
-  test('creates a Studio and Next.js frontend with their scoped env files', async () => {
+  test('creates a Studio and Next.js frontend with identical env files', async () => {
     await expect(scaffoldProject(options)).resolves.toMatchObject({
       frontendDependenciesInstalled: true,
-      frontendEnv: {
-        NEXT_PUBLIC_SANITY_DATASET: 'production',
-        NEXT_PUBLIC_SANITY_PROJECT_ID: 'abc123',
-      },
-      frontendEnvPrefix: 'NEXT_PUBLIC_',
-      frontendEnvWritten: true,
       frontendPackageManager: 'npm',
       frontendPath: path.join(options.workDir, 'web'),
-      studioEnvWritten: true,
       studioPath: path.join(options.workDir, 'sanity'),
     })
 
@@ -119,63 +114,30 @@ describe('scaffoldProject', () => {
       expect.objectContaining({dirName: 'web', packageManager: 'npm'}),
     )
     expect(mockInstallFrontendDeps).toHaveBeenCalled()
-    expect(mockAppendEnvValues).toHaveBeenCalledWith(
+    expect(mockWriteFileSync).toHaveBeenCalledWith(
       path.join(options.workDir, 'sanity', '.env.local'),
-      {SANITY_AUTH_TOKEN: 'sk-token'},
-      expect.anything(),
+      envContents,
+      'utf8',
     )
-    expect(mockAppendEnvValues).toHaveBeenCalledWith(
+    expect(mockWriteFileSync).toHaveBeenCalledWith(
       path.join(options.workDir, 'web', '.env.local'),
-      {
-        NEXT_PUBLIC_SANITY_DATASET: 'production',
-        NEXT_PUBLIC_SANITY_PROJECT_ID: 'abc123',
-      },
-      expect.anything(),
+      envContents,
+      'utf8',
     )
   })
 
-  test.each([
-    ['nuxtjs', FRONTEND_ENV_PREFIX_OVERRIDES.nuxtjs],
-    ['sveltekit-1', FRONTEND_ENV_PREFIX_OVERRIDES['sveltekit-1']],
-  ])('uses framework-correct public env keys for %s', async (slug, prefix) => {
-    const onFrameworkDetected = vi.fn()
+  test('leaves a detected frontend unchanged', async () => {
     mockDetectFrameworkRecord.mockResolvedValue({
-      envPrefix: 'IGNORED_',
-      name: slug,
-      slug,
-    })
-
-    await expect(scaffoldProject({...options, onFrameworkDetected})).resolves.toMatchObject({
-      detectedFramework: slug,
-      frontendEnv: {
-        [`${prefix}SANITY_DATASET`]: 'production',
-        [`${prefix}SANITY_PROJECT_ID`]: 'abc123',
-      },
-      frontendEnvWritten: false,
-      studioEnvWritten: true,
-    })
-    expect(onFrameworkDetected).toHaveBeenCalledWith(slug)
-    expect(mockCreateFrontend).not.toHaveBeenCalled()
-  })
-
-  test('uses Astro public env keys from framework metadata', async () => {
-    mockDetectFrameworkRecord.mockResolvedValue({
-      envPrefix: 'PUBLIC_',
-      name: 'Astro',
-      slug: 'astro',
+      name: 'Nuxt',
+      slug: 'nuxtjs',
     })
 
     await expect(scaffoldProject(options)).resolves.toMatchObject({
-      detectedFramework: 'Astro',
-      frontendEnv: {
-        PUBLIC_SANITY_DATASET: 'production',
-        PUBLIC_SANITY_PROJECT_ID: 'abc123',
-      },
-      frontendEnvPrefix: 'PUBLIC_',
-      frontendEnvWritten: false,
-      studioEnvWritten: true,
+      detectedFramework: 'Nuxt',
+      studioPath: path.join(options.workDir, 'sanity'),
     })
     expect(mockCreateFrontend).not.toHaveBeenCalled()
+    expect(mockWriteFileSync).toHaveBeenCalledTimes(1)
   })
 
   test('returns a typed recovery when frontend creation fails', async () => {
@@ -184,28 +146,44 @@ describe('scaffoldProject', () => {
 
     await expect(scaffoldProject(options)).resolves.toMatchObject({
       frontendCreationError: 'create-next-app failed',
-      frontendEnvWritten: false,
-      studioEnvWritten: true,
+      frontendPackageManager: 'npm',
     })
+    expect(mockWriteFileSync).toHaveBeenCalledTimes(1)
   })
 
-  test('returns typed env and dependency recovery states', async () => {
+  test('returns a typed dependency recovery state', async () => {
     const {FrontendScaffoldError} = await import('../createFrontend.js')
-    mockAppendEnvValues
-      .mockReturnValueOnce({created: false, skippedKeys: ['SANITY_AUTH_TOKEN'], wroteKeys: []})
-      .mockReturnValueOnce({
-        created: false,
-        skippedKeys: ['NEXT_PUBLIC_SANITY_PROJECT_ID'],
-        wroteKeys: [],
-      })
     mockInstallFrontendDeps.mockRejectedValue(new FrontendScaffoldError('next-sanity failed'))
 
     await expect(scaffoldProject(options)).resolves.toMatchObject({
       frontendDependenciesInstalled: false,
       frontendDependencyError: 'next-sanity failed',
-      frontendEnvWritten: false,
-      studioEnvWritten: false,
     })
+  })
+
+  test('reports a Studio env write failure after Studio scaffolding', async () => {
+    mockWriteFileSync.mockImplementationOnce(() => {
+      throw new Error('read-only filesystem')
+    })
+
+    await expect(scaffoldProject(options)).rejects.toThrow(
+      'Studio scaffold completed, but writing ./sanity/.env.local failed: read-only filesystem',
+    )
+    expect(mockCreateFrontend).not.toHaveBeenCalled()
+  })
+
+  test('reports a website env write failure after website scaffolding', async () => {
+    mockWriteFileSync
+      .mockImplementationOnce(() => undefined)
+      .mockImplementationOnce(() => {
+        throw new Error('read-only filesystem')
+      })
+
+    await expect(scaffoldProject(options)).rejects.toThrow(
+      'Website scaffold completed, but writing ./web/.env.local failed: read-only filesystem',
+    )
+    expect(mockCreateFrontend).toHaveBeenCalled()
+    expect(mockInstallFrontendDeps).not.toHaveBeenCalled()
   })
 
   test('reports next-sanity as pending when package installation is manual', async () => {
